% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fucntionsmtg.R
\name{calcmtg}
\alias{calcmtg}
\title{Generating parameters}
\usage{
calcmtg(mu, si, rho, spot, r, dt)
}
\arguments{
\item{mu}{average historical return}

\item{si}{sd of historical returns}

\item{rho}{covariance matrix (diagonal=1) of historical returns}

\item{spot}{vector of spot prices}

\item{r}{risk-free interest rate}

\item{dt}{the time (in days) between checking prices}
}
\value{
Returns list containing: mean of historical returns, sd of historical returns, covariance matrix (1 on diagonal),
spot prices, mean of returns for mtg measure, covariance matrix of returns in mtg measure, R-N derivative, risk-free interest rate
}
\description{
Generates parameters of martingale measure from historical parameters
Mostly used in estim function.
}
