% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fucntionsmtg.R
\name{estim}
\alias{estim}
\title{Generating parameters}
\usage{
estim(pricedata, r, t = 0, dt = 1)
}
\arguments{
\item{pricedata}{matrix of historical prices (columns~assets)}

\item{r}{risk-free interest rate}

\item{t}{the time (in days) from which the parameters are computed}

\item{dt}{the time (in days) between checking prices}
}
\value{
Returns list containing: mean of historical returns, sd of historical returns, covariance matrix (1 on diagonal),
spot prices, mean of returns for mtg measure, covariance matrix of returns in mtg measure, R-N derivative, risk-free interest rate
}
\description{
Generates parameters of martingale measure from historical data
}
\examples{
data(wig)
par <- estim(wig_m,0.02)
par2<- estim(wigkgh_m,0.0135)
}
