% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fucntionsmtg.R
\name{gentraj}
\alias{gentraj}
\title{Generating price trajectory}
\usage{
gentraj(par, t, dt, method = 1)
}
\arguments{
\item{par}{parameters in format as result of the estim function}

\item{t}{the time (in days) of generated trajectory}

\item{dt}{the time (in days) between checking prices}

\item{method}{1 - computes price using martingale measure directly, 2 - computes prices from original measure and Radon-Nikodym derivative}
}
\value{
For method 1 returns matrix of prices (columns~assets), for method 2 returns list of price matrix and R-N derivative
}
\description{
Generates random price trajectory for a given time period
}
\examples{
rn <- function(x) 1
p <- list(-0.000228,0.00956,matrix(1,1,1),2204,0,matrix(0.000108,1,1),rn,0.02)
tr <- gentraj(p,200,1)
}
