% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fucntionsmtg.R
\name{payapply}
\alias{payapply}
\title{Estimating option price}
\usage{
payapply(par, t, dt = 1, method = 1, payoff, assets = 1, N = 10000)
}
\arguments{
\item{par}{parameters in format as result of the estim function}

\item{t}{the time (in days) of generated trajectory}

\item{dt}{the time (in days) between checking prices}

\item{method}{1 - computes price using martingale measure directly, 2 - computes prices from original measure and Radon-Nikodym derivative}

\item{payoff}{function: trajectory |-> option payoff}

\item{assets}{vector of assets to which trajectories should be applied payoff function}

\item{N}{number of repeats}
}
\value{
Returns vector of option prices at t=0
}
\description{
Computes average option payoff
}
\examples{
rn <- function(x) 1
p <- list(-0.000228,0.00956,matrix(1,1,1),2204,0,matrix(0.000108,1,1),rn,0.02)
payoff2200 <- function(x) max(0,x[length(x)]-2200)
prices <- payapply(p,200,payoff=payoff2200,N=20000)
}
